Managing the Cohort object

The list of methods designed for managing the Cohort configuration and state.

add_source - Add source to Cohort object.

update_source - Update Cohort object source.

add_step - Add step to Cohort object.

rm_step - Remove step from Cohort object.

add_filter - Add filter to Cohort step.

rm_filter - Remove filter from Cohort step.

update_filter - Update filter configuration.

run - Run data filtering.

Value

The object of class `Cohort` having the modified configuration dependent on the used method.